Factor tt::is_streamable into its own file
